v0.00.8
Created a system for save files in which if a save file is imported its index is stored and then when user exists the app, the old data at that index is overwritten with the new data. Basically, stores user's progress.

Created overWriteFile function to overwrite old data in save_file.txt